+++ tests/test_charts.py
@@ -63,13 +63,13 @@ class TestFlowDuration(unittest.TestCase):
 class TestCyclePlot(unittest.TestCase):
 
     def test_charts_cycleplot_exists(self):
-        expected = hf.extract_nwis_df(test_json)
+        expected = hf.extract_nwis_df(test_json, interpolate=False)
         actual_fig, actual_ax = charts.cycleplot(expected)
         self.assertIsInstance(actual_fig, matplotlib.figure.Figure)
         self.assertIsInstance(actual_ax[0], matplotlib.axes.Axes)
 
     def test_charts_cycleplot_parts(self):
-        expected = hf.extract_nwis_df(test_json)
+        expected = hf.extract_nwis_df(test_json, interpolate=False)
 
         actual_fig, actual_ax = charts.cycleplot(expected)