+++ tests/test_exceptions.py
@@ -6,6 +6,7 @@ test_exceptions.py
 """
 from __future__ import absolute_import, print_function, division, unicode_literals
 import unittest
+import warnings
 
 from hydrofunctions import exceptions
 
@@ -18,6 +19,22 @@ class TestExceptions(unittest.TestCase):
 
     def test_exceptions_HydroNoDataError_can_be_raised(self):
         self.assertRaises(exceptions.HydroNoDataError, raiseHydroNoDataError)
+
+    def test_exceptions_HydroNoDataError_can_be_caught(self):
+        actual = False
+        try:
+            raiseHydroNoDataError()
+        except exceptions.HydroNoDataError as err:
+            actual = True
+        self.assertTrue(actual, "The HydroNoDataError should have been caught, but wasn't.")
+
+
+class TestWarnings(unittest.TestCase):
+
+    def test_exceptions_HydroUserWarning_can_be_called(self):
+        with self.assertWarns(exceptions.HydroUserWarning):
+            warnings.warn("test warning message", exceptions.HydroUserWarning)
+
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)